+++ src/pages/senimarAttend/components/BossCardComponent.tsx
@@ -3,6 +3,7 @@ import { ThemeProvider, Select, Option } from '@material-tailwind/react';
 import Countdown from './Countdown';
 import SeminarButton from './SeminarButton';
 import SeminarSelector from './SeminarSelector';
+import SeminarInput from './SeminarInput';
 
 const BossCardComponent = () => {
   const [attendValue, setAttendValue] = useState<number>(5);
@@ -38,24 +39,9 @@ const BossCardComponent = () => {
     },
   };
 
-  const inputStyle: React.CSSProperties = {
-    textAlign: 'center',
-    fontSize: '28px',
-    fontWeight: '600',
-    height: '52px',
-    maxWidth: '42px',
-    color: 'black',
-    background: 'white',
-  };
-
   return (
     <ThemeProvider value={selectTheme}>
-      <div className="mx-auto my-[22px] flex h-[52px] w-[192px] justify-between">
-        <input disabled style={inputStyle} />
-        <input disabled style={inputStyle} />
-        <input disabled style={inputStyle} />
-        <input disabled style={inputStyle} />
-      </div>
+      <SeminarInput disabled />
       <SeminarButton onClick={startSeminar}>시작</SeminarButton>
       {seminarExist ? (
         <Countdown />
