+++ src/pages/senimarAttend/components/MemberCardComponent.tsx
@@ -2,6 +2,7 @@ import React, { useState } from 'react';
 import { Typography } from '@mui/material';
 import Countdown from './Countdown';
 import SeminarButton from './SeminarButton';
+import SeminarInput from './SeminarInput';
 
 const inputStyle: React.CSSProperties = {
   textAlign: 'center',
@@ -13,17 +14,12 @@ const inputStyle: React.CSSProperties = {
 };
 
 const MemberCardComponent = () => {
-  const inputListKey = [0, 1, 2, 3]; // key값 어떻게 줘야할지 모르겠어서 이렇게 했습니다 좋은방법좀ㅠㅠ
   const [componentIncorrectAlert, setComponentIncorrectAlert] = useState(<p className="mb-[22px]" />);
   const [seminarExist, setSeminarExist] = useState(true); // TODO: api로 교체
 
   return seminarExist ? (
     <>
-      <div className="mx-auto mt-[22px] flex h-[52px] w-[192px] justify-between">
-        {inputListKey.map((key) => {
-          return <input key={key} style={inputStyle} maxLength={1} />;
-        })}
-      </div>
+      <SeminarInput disabled={false} />
       <div className="mx-auto flex items-center justify-center text-small text-red-500">{componentIncorrectAlert}</div>
       <SeminarButton>출석</SeminarButton>
       <Countdown />
